Add a trigger character based autocompletion extension
This is quite minimal, all it does is trigger the omnifunc if any of the
trigger characters is typed in insert mode.

The dap omnifunc implementation is async, so this is non-blocking and
should be fast.

This should provide a good experience without dap turning into a full
blown autocompletion plugin.