Add listeners.{before,after} as replacement for custom_ handlers
As discussed in https://github.com/mfussenegger/nvim-dap/pull/124 this
gives users the possible to choose whether the listeners should be
called before or after the internal listeners.

I also used the opportunity to add another change: The response
listeners are now also called in error cases. So the number of arguments
changed.

`custom_event_handlers` and `custom_response_handlers` are both still
there and used. (Without the signature change for request responses)
If used a warning will be printed out.

They'll be removed at some point in the future.