Keep repl.append content if repl is hidden
During debugging (or when running junit tests with extensions like
nvim-jdtls) output may be written to the REPL. Previously this output
was discarded if the REPL wasn't open.

This commit changes the behavior to preserve the output until the next
debug session starts - or until the REPL was opened and is then closed
again.