Improve config filetype/adapter mapping on launch.json reads
In a configuration, the `type` field refers to an adapter defined in
`dap.adapters.<adapter-name>`. The current implementation assumes that
the `config.type`, which is the name of an adapter, is the same as the
filetype.

The assumption works in most cases, since most adapter names match the
filetype. But the assumption breaks with adapters like lldb, which fits
several filetypes (c, c++, rust, and others).

This commit tries to infer the file types based on existing
configurations.