Allow to override the active session
This adds a (as of now undocumented) `set_session` method that allows
extensions to change the active session with which the user facing
commands are interacting.

This is useful for extensions like nvim-dap-python because debug
adapters like debugpy contain custom extensions to the core debug
adapter protocol where they need to have multiple sessions running at
the same time.

Until multi-sessions become part of the official specification,
multi-session support is out-of-scope for nvim-dap, but this method
should be enough to allow extensions to do the remaining work.

(The UX may not be ideal because breakpoints are global and there exists
no breakpoint - session mapping, but it should at least make it possible
to get things working in some way)