Avoid repl lazy loading to avoid crash on latest master
There's currently an issue on neovim master that causes loading
(uncached) modules within a luv callback to crash neovim.

This moves the `require(dap.repl)` call to the top of the session - the
session itself is already lazy loaded and it should be fine to pull in
all dependencies in that case.