Ensure ${file} maps to absolute path and {$relativeFile} to relative path
Both breaking change and bugfix, see
https://github.com/mfussenegger/nvim-dap/issues/351 for context.

Without this change, the value of `${file}` depends on how you open a
file. `:e myFile` will resolve to `myFile` and `:e /path/to/myFile`
resolves to `/path/to/myFile`.

After this change `${file}` will always return the absolute path,
independent of how you open a file.

Side effect of this is that it breaks usages of
`${workspaceFolder}/${file}`.