Close session on disconnect response
Closes https://github.com/mfussenegger/nvim-dap/issues/475
Closes https://github.com/mfussenegger/nvim-dap/issues/469

A consequence of this change is that even for the debug adapters that do
send a terminated event after receiving a disconnect request (like
java-debug), this event might not be received anymore because the
session is closed after the disconnect response.

Extensions that want to run some logic if a debug session ends
should listen to both `event_terminated` and `disconnect`.