Support command:pickRemoteProcess
The little documented pickRemoteProcess command can be used with the
pipeTransport option to query processes running on a potentially remote
host rather than the local machine.

This effectively works by running the command to list processes in the
same environment that the debugger will be run in.

See https://code.visualstudio.com/docs/cpp/pipe-transport for details.