Increase chance to trigger request callbacks on debug adapter exit
If the client sends a terminate the debug adapter can send the response,
terminate event and terminate the debug adapter process. In that case
the client immediately closed the stdout handle and cleared the message
callbacks table, preventing it from reliably processing the last few
messages.

This makes use of the async stdout:shutdown/client:shutdown to give it a
chance to process these last messages.

This is a follow up to
https://github.com/mfussenegger/nvim-dap/commit/03807a4682081cc2523219ef17a36a618995b36d
and should ensure that event_terminated listeners triggers happen in
most (all?) cases.