Add support for LuaRocks + release workflow
nvim-dap is a common dependency across Neovim plugins.
Using luarocks may alleviate the need for users to specify their
plugins' dependencies in their plugin manager
(e.g., vim-plug or packer). See also
[this blog post](https://teto.github.io/posts/2021-09-17-neovim-plugin-luarocks.html)
for details.

This PR adds a release workflow that publishes this plugin to
[LuaRocks](https://luarocks.org/) whenever a tag is pushed,
as well as a rockspec that can be used to release to the `dev` channel.

For the release workflow to work, someone with a LuaRocks account
will have to add their API key to this repo's secrets.

Note that I have added a shield to the readme that assumes the
existence of an `mfussenegger/nvim-dap` LuaRocks module.