Add a new option dap.repl.show_tree: whether to show the tree in REPL
A new option `show_tree` can be set by users to configure whether to
show the expandable tree in the DAP REPL window. Concise and simple
representation in the REPL would be useful for many languages,
see #737 for an example (python).

The option can be configured using, for instance:

```
require('dap.repl').show_tree = false
```