feat(vscode): implement `run_launchjs` to run config interactively
Although `DapLoadLaunchJSON` already exists, it does not provide an easy
to use and interactive way of running configurations. For instance, if
you'd wanted to load a `codelldb` config for a `cpp` file, you would
have had to execute the underlying lua API (`load_launchjs`) manually
and pass `cpp` as `type_to_filetypes` parameter.