Fix repl expressions leave an empty prompt in the middle (#1063)
Fixes #1052.

Explanation:

* REPL buffer is based on `buftype=prompt`, where nvim makes sure that
there is always a prompt at the end of buffer in insert mode (`dap> ` ).
* When a new command is entered, `exec` function in `repl.lua` is
executed (registered via `vim.fn.prompt_setcallback`). nvim does not put
the next prompt until callback returns, so, within the callback, new
lines can simply be appendend at the end of the buffer. (e.g. `.help`
command does not leave an empty prompt in the middle).
* Things get a bit tricky with expression command because when a DAP
server request comes back, nvim-dap is aldready outside of the callback,
and there is an empty prompt at the end of the buffer. Two options here
so not to leave an empty prompt in the middle,
  1. insert the new line(s) right in front of the empty prompt
  2. overwrite the last line with the new line(s)
* `append` function inside `repl.lua` is updated to apply the first
option. `evaluate_handler` now calls this function instead of
`layer.render` where possible.
* The second opton is used where `tree.render` is called inside
`evaluate_handler`.