Ensure dap.status() shows start progress
nvim-dap reports `Launching debug adapter` before evaluating a adapter
function:

    progress.report('Launching debug adapter')

But that message was never shown in `dap.status()` because at that point
no session is active and `status()` returned a `''`.

If the adapter function takes a long time, it gave
the impression that nothing was happening.

Solution:

Always show the last message in `status()`, and send a `""` if a session
closes.