Restore stopped state if goto errors
Problem:

A debug adapter can respond to goto with an error - e.g. if attempting
to jump to a location where there is no code to execute.

If that happened, selecting another goto location was no longer possible
because the stopped state had been cleared.

Solution:

Restore stopped state if goto fails