Add an action for variables to copy expression's to unnamed register
It's often the case that you're drilling into a list of objects and then
want to inspect some object in more detail - maybe run some methods on
it interactively or something.

This adds a `Copy as expression` item to the actions menu for variables
if the debug adapter had provided an `evaluateName` (mapped to `o` by default).
If picked it is put into the unnamed register so you can use it in the
REPL to further inspect the variable in question.

E.g. if you drill into `dict -> data -> 24 -> country -> name` and use
`Copy as expression` on `name` you'd get
`dict['data']['24']['country']['name']` with `debugpy`