Fix :ga from within REPL
`:ga` Needed to be executed twice if executed from within the REPL
buffer due to a `nvim_buf_delete`/`nvim_win_close` running into a
"Cannot close last window" error.

This replaces the close call with a `modified` change, which serves the
purpose of not blocking an exit if the REPL prompt buffer is modified.